feat：add get category percent group list interface

+++ src/mbill.Service/Bill/Bill/Output/CategoryPercentGroupDto.cs
@@ -15,7 +15,7 @@ public class CategoryPercentItemDto
 
     public string CategoryIcon { get; set; }
 
-    public float Percent { get; set; }
+    public double Percent { get; set; }
 
     public string Amount { get; set; }
 }
